Split coordinator binary into lib.rs and main.rs
Allows using the `run` function from tests and examples.
